refactor: 패키지 네이밍 model -> domain/model
