+++ app/src/main/java/woowacourse/payments/domain/model/CardInfo.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.model
+package woowacourse.payments.domain.model
 
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
