+++ app/src/main/java/woowacourse/payments/ui/addcard/component/VendorItem.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import woowacourse.payments.R
-import woowacourse.payments.model.CardVendor
+import woowacourse.payments.domain.model.CardVendor
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
 import woowacourse.payments.ui.uimodel.VendorUiModel
 import woowacourse.payments.ui.uimodel.toUiModel
