+++ app/src/main/java/woowacourse/payments/ui/addcard/model/VendorModalUiModel.kt
@@ -2,7 +2,7 @@ package woowacourse.payments.ui.addcard.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
-import woowacourse.payments.model.CardVendor
+import woowacourse.payments.domain.model.CardVendor
 import woowacourse.payments.ui.uimodel.VendorUiModel
 import woowacourse.payments.ui.uimodel.toUiModel
